path_to_enlightenment.py:  Added _useful_ tests.
2 new TestCase classes: TestFilterKoanNames and TestKoansSuite.

TestKoansSuite needs more and better tests, but TestCase and TestSuite
objects are hard to test without importing every constituent test by
hand.  That would require hard-coding the koans into the tests, which
would break the instant someone adds a new koan... which is what
loading the koans from a config file is supposed to fix in the first
place.